+++ src/Users/UsersService.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Laravel\Forge\Users;
+
+use Laravel\Forge\Users\Commands\GetAuthenticatedUserCommand;
+
+class UsersService
+{
+    public function get()
+    {
+        return (new GetAuthenticatedUserCommand());
+    }
+}